# improving performance of PuppiProducer