--- a/CommonTools/PileupAlgos/src/PuppiCandidate.cc
+++ b/CommonTools/PileupAlgos/src/PuppiCandidate.cc
@@ -1,1 +1,0 @@
-#include "CommonTools/PileupAlgos/interface/PuppiCandidate.h"