#### PR description:

This PR is an attempt to improve the performance (in terms of execution time) of the `PuppiProducer` plugin.

No changes in the outputs are expected.

Most of these technical changes are not in the producer itself, but in the `PuppiContainer` class; all changes are related (directly or indirectly) to the function `var_within_R` (the "alpha" variable on which the PUPPI weight is based), which appears to be the most time-consuming part of the algorithm.

The table below summarizes a few timing estimates for the `puppi` module in the RECO step for a Run-3 and Phase-2 workflow.

| wf | time pre-PR [ms] | time post-PR [ms] |
|---:|---:|---:|
| Run-3 RECO [(QCD MC)](https://cmsweb.cern.ch/das/request?input=dataset%3D%2FRelValQCD_FlatPt_15_3000HS_14%2FCMSSW_11_2_0_pre3-PU_112X_mcRun3_2021_realistic_v5-v1%2FGEN-SIM-DIGI-RAW&instance=prod/global)    |  74.2 |  33.2 |
| Phase-2 RECO [(QCD MC)](https://cmsweb.cern.ch/das/request?input=dataset%3D%2FRelValQCD_Pt15To7000_Flat_14%2FCMSSW_11_2_0_pre3-PU25ns_110X_mcRun4_realistic_v3_2026D49PU200_rsb-v1%2FGEN-SIM-DIGI-RAW&instance=prod/global) |  620.1 | 255.9  |

The numbers are taken from the `FastTimerService`, running on 100 events; they suggest that the overall reduction in timing is approx. 55-60%.

Here is a short summary of the changes in the PR (below, `goodVar` and `var_within_R` are used interchangeably, since the former function is simply a call to the latter):

* `PuppiCandidate` is converted to a simple `struct`, removing its inheritance from `fastjet::PseudoJet` (looking at the PUPPI implementation as a whole, the usage of `fastjet` does not seem necessary); this change accounts for most of the speedup (approx. 35-40% speedup);

* skip the calculation of `goodVar` when not needed, e.g. for candidates that are assigned a default weight (0, or 1); this seems to bring roughly a 10-15% improvement in performance.

* another bit of speedup is obtained by taking one condition (`puppi_id != 3`) out of `var_within_R`, and using it just once in `PuppiContainer::initialize` to fill the collections later used as inputs to `goodVar`; this leads to adding one more data member (a `vector` of `PuppiCandidate`s) to the `PuppiContainer` class.

In this PR, there are two spots where things could certainly be improved (but improving them will lead to small changes in the outputs):

* in `PuppiContainer::initialize`, an instance of `fastjet::PseudoJet` ([here](https://github.com/cms-sw/cmssw/compare/master...missirol:devel_112X_puppi_speedup2?expand=1#diff-920ad87cd21790b241508fda9a732044R49)) is used to obtain the 4-momenta of the candidates used by PUPPI; this was the only way I could find to maintain the same numbers as pre-PR (this may be due to the fact that the inputs used for `reset_PtYPhiM` are floats, and then `fastjet` recomputes and stores the p4 using doubles); one way to simplify this (assuming the differences in the outputs are ultimately not significant) would be to remove `PuppiCandidate` altogether and simply use the `RecoObj` class already in use in the `PuppiProducer` (maybe, changing the floats in `RecoObj` to doubles).

* in `goodVar`, the `DeltaR2` between two given candidates is calculated (at most) twice ([here](https://github.com/cms-sw/cmssw/compare/master...missirol:devel_112X_puppi_speedup2?expand=1#diff-920ad87cd21790b241508fda9a732044R96)), first with rapidity, then with pseudo-rapidity; this is not modified by this PR (otherwise, the outputs would inevitably change), but maybe it’s something that could be simplified.

FYI: @ahinzmann @lathomas @kirschen

#### PR validation:

Checked that the PUPPI weights of all PF candidates are unchanged for 200 events (100 for a Run-3 workflow, and 100 for a Phase-2 workflow), for both `puppi` (used for jets) and `puppiNoLep` (used for MET).
Standard workflows, i.e. `runTheMatrix.py -l limited -i all --ibeos`, ran successfully.

#### if this PR is a backport please specify the original PR and why you need to backport that PR:

N/A