@slava77 2020-08-18
@mrodozov @smuzaffar 
do we have some problems with webhooks?
this one was posted for 14 minutes already (at the time of my message)