@mrodozov 2020-08-18
@slava77 the remote calls to jenkins hanged with proxy error 502 and I'm trying to restart the service but I don't have sudo rights to do it