@davidlange6 2020-08-18
It seems there are other problems at cern today (at least affecting cms computing infrastructure)

On Aug 18, 2020, at 3:33 PM, Mircho Rodozov <notifications@github.com<mailto:notifications@github.com>> wrote:



@slava77<https://github.com/slava77> the remote calls to jenkins hanged with proxy error 502 and I'm trying to restart the service but I don't have sudo rights to do it

—
You are receiving this because you are subscribed to this thread.
Reply to this email directly, view it on GitHub<https://github.com/cms-sw/cmssw/pull/31164#issuecomment-675481484>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/ABGPFQ2VYBUIGTQZJX5IVY3SBJ7LRANCNFSM4QDNGPKA>.