@mrodozov 2020-08-18
@davidlange6 do you by any chance happen to have root permissions for the jenkins03 machine ?