@davidlange6 2020-08-18
No, I don’t - perhaps ask IT to reboot it?

On Aug 18, 2020, at 4:45 PM, Mircho Rodozov <notifications@github.com<mailto:notifications@github.com>> wrote:



@davidlange6<https://github.com/davidlange6> do you by any chance happen to have root permissions for the jenkins03 machine ?

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub<https://github.com/cms-sw/cmssw/pull/31164#issuecomment-675522818>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/ABGPFQ5R5XJMARLSV773IOLSBKHZ3ANCNFSM4QDNGPKA>.