@cmsbuild 2020-08-18
+1
Tested at: 0c3b885fbbd8e493f396d89bcdc0687cf52f0818
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-63bdad/8799/summary.html
CMSSW: CMSSW_11_2_X_2020-08-17-2300
SCRAM_ARCH: slc7_amd64_gcc820