@slava77 2020-08-24
+1

for https://github.com/cms-sw/cmssw/pull/31164	e0f02d3 
- code changes are in line with the PR description and the follow up review
    - the last commits removed dependence on fastjet, but introduced a numerical difference https://github.com/cms-sw/cmssw/pull/31164/files/0c3b885fbbd8e493f396d89bcdc0687cf52f0818..e0f02d36a5b2d0fd5c5fbe53358cbb902f1700cf in the way how rapidity/pseudorapidity are set, something to be noted/bypassed in a possible backport
- jenkins tests pass and comparisons with the baseline show small differences in puppi-related variables, consistent with the change in the numerical precision of the kinematics
-  

curiously enough, there was no actual overlap in the context lines with #31174