@missirol 2020-08-24
> the last commits removed dependence on fastjet, but introduced a numerical difference

@slava77 

In case it helps: numerical changes in the outputs are introduced only by the very last commit of the PR, i.e. e0f02d36a5b2
(we didn't verify this explicitly in the PR review, but I had checked it locally).