@missirol 2020-08-25
@slava77 , one late question from my side: would it be okay to backport this (except for the last commit) to `11_1_X`, so that it can be used for the Phase-2 HLT TDR studies?