@kpedro88 2020-08-26
@missirol thanks for this PR! It's a worthy followup to my initial effort in #23270.

Could you make the same limited backport to 10_6_X? It would be nice to propagate this speedup to users for ultra-legacy analysis.