@slava77 2020-08-27
> I'm looking into a backport of this PR to `10_6_X` (taking into account the no-change policy). It would not be entirely trivial, because the Puppi implementation in earlier releases (incl. `10_6`) is significantly different; nevertheless, I'll give it a try.

Thank you.